fix: split resolveHome into platform files, warn on UserHomeDir failure, panic on bad embedded config, add paths package doc
